+++ app/models/user.rb
@@ -13,13 +13,10 @@ class User
   field :per_page, :type => Fixnum, :default => PER_PAGE
   field :time_zone, :default => "UTC"
 
-  after_destroy :destroy_watchers
   before_save :ensure_authentication_token
 
   validates_presence_of :name
   validates_uniqueness_of :github_login, :allow_nil => true
-
-  attr_protected :admin
 
   has_many :apps, :foreign_key => 'watchers.user_id'
 
@@ -59,10 +56,5 @@ class User
     self[:github_login] = login
   end
 
-  protected
-
-    def destroy_watchers
-      watchers.each(&:destroy)
-    end
 end